+++ apps/wrapper/applicant-portal/src/pages/CreateForm.jsx
@@ -20,7 +20,8 @@ import {
 } from "./../forms";
 
 import APPLICANT_ROUTE_MAP from "../routes/ApplicantRoute";
-import { Card } from "../components";
+import { setCookie,  } from "../utils";
+import { Button, Card } from "../components";
 import CommonModal from "../Modal";
 import Toast from "../components/Toast";
 import "./loading.css";
@@ -32,6 +33,8 @@ import {
   getFormURI,
   updateFormSubmission,
 } from "../api/formApi";
+import { generate_uuidv4 } from "../utils";
+import { applicantService } from "../services";
 
 const ENKETO_URL = process.env.REACT_APP_ENKETO_URL;
 
@@ -40,6 +43,7 @@ const CreateForm = (props) => {
 
   let { formName, formId, applicantStatus } = useParams();
   let [encodedFormURI, setEncodedFormURI] = useState("");
+  let [paymentDetails,setPaymentDetails] = useState("")
   let [onFormSuccessData, setOnFormSuccessData] = useState(undefined);
   let [formDataNoc, setFormDataNoc] = useState({});
   let [onFormFailureData, setOnFormFailureData] = useState(undefined);
@@ -113,7 +117,8 @@ const CreateForm = (props) => {
       if (formId) {
         const postData = { form_id: formId };
         const res = await getFormData(postData);
-        formData = res.data.form_submissions[0];
+        formData = res?.data?.form_submissions[0];
+        setPaymentDetails(formData?.payment_status)
         setFormDataNoc(formData);
       }
     }
@@ -305,6 +310,47 @@ const CreateForm = (props) => {
     navigate(`${APPLICANT_ROUTE_MAP.dashboardModule.my_applications}`);
   };
 
+  const handlePayment = async () => {
+    // setcookies here
+    setCookie(
+      "formId",formId
+    );
+    const instituteDetails = getCookie("institutes");
+    const instituteId = instituteDetails?.[0]?.id;
+    const postData = {
+      endpoint: "https://eazypayuat.icicibank.com/EazyPG",
+      returnUrl: "https://payment.uphrh.in/api/v1/user/payment",
+      paymode: "9",
+      secret: "",
+      merchantId: "600547",
+      mandatoryFields: {
+        referenceNo: generate_uuidv4(),
+        submerchantId: "45",
+        transactionAmount: "10",
+        invoiceId: "x1",
+        invoiceDate: "x",
+        invoiceTime: "x",
+        merchantId: "x",
+        payerType: "affiliation",
+        payerId: instituteId,
+        transactionId: "x",
+        transactionDate: "x",
+        transactionTime: "x",
+        transactionStatus: "x",
+        refundId: "x",
+        refundDate: "x",
+        refundTime: "x",
+        refundStatus: "x",
+      },
+      optionalFields: "",
+    };
+    try {
+      const paymentRes = await applicantService.initiatePayment(postData);
+      await window.open(paymentRes?.data?.redirectUrl);
+    } catch (error) {}
+  };
+
+
   const handleFormDownload = async () => {
     try {
       setIsDownloading(true);
@@ -427,6 +473,16 @@ const CreateForm = (props) => {
             >
               Download NOC/Certificate
             </button>
+            <button
+          className={`${
+            paymentDetails === "Pending"
+              ? "border border-blue-900 bg-blue-900 text-white rounded-[4px] px-2 h-[44px]"
+              : "cursor-not-allowed border border-gray-500 bg-white rounded-[4px] text-gray-200 px-2 h-[44px]"
+          }`}
+          disabled={paymentDetails === "Pending" ? false : true}
+          onClick={handlePayment}
+          
+        >Pay</button>
           </div>
         </div>
